Software/requirements.txt
# Packaged installed via pip:
pyserial
tqdm
PyQt6
requests
